+++ mongodb_exporter.go
@@ -53,29 +53,7 @@ var (
 		Default("mongodb://localhost:27017").
 		Envar("MONGODB_URI").
 		String()
-
-	authDB   = kingpin.Flag("mongodb.authentification-database", "Specifies the database in which the user is created").Default("").String()
-	tlsF     = kingpin.Flag("mongodb.tls", "Enable tls connection with mongo server").Bool()
-	tlsCertF = kingpin.Flag("mongodb.tls-cert", "Path to PEM file that contains the certificate (and optionally also the decrypted private key in PEM format).\n"+
-		"    \tThis should include the whole certificate chain.\n"+
-		"    \tIf provided: The connection will be opened via TLS to the MongoDB server.").Default("").String()
-	tlsPrivateKeyF = kingpin.Flag("mongodb.tls-private-key", "Path to PEM file that contains the decrypted private key (if not contained in mongodb.tls-cert file).").Default("").String()
-	tlsCAF         = kingpin.Flag("mongodb.tls-ca", "Path to PEM file that contains the CAs that are trusted for server connections.\n"+
-		"    \tIf provided: MongoDB servers connecting to should present a certificate signed by one of this CAs.\n"+
-		"    \tIf not provided: System default CAs are used.").Default("").String()
-	tlsDisableHostnameValidationF = kingpin.Flag("mongodb.tls-disable-hostname-validation", "Disable hostname validation for server connection.").Bool()
-	maxConnectionsF               = kingpin.Flag("mongodb.max-connections", "Max number of pooled connections to the database.").Default("1").Int()
-	testF                         = kingpin.Flag("test", "Check MongoDB connection, print buildInfo() information and exit.").Bool()
-
-	socketTimeoutF = kingpin.Flag("mongodb.socket-timeout", "Amount of time to wait for a non-responding socket to the database before it is forcefully closed.\n"+
-		"    \tValid time units are 'ns', 'us' (or 'µs'), 'ms', 's', 'm', 'h'.").Default("3s").Duration()
-	syncTimeoutF = kingpin.Flag("mongodb.sync-timeout", "Amount of time an operation with this session will wait before returning an error in case\n"+
-		"    \ta connection to a usable server can't be established.\n"+
-		"    \tValid time units are 'ns', 'us' (or 'µs'), 'ms', 's', 'm', 'h'.").Default("1m").Duration()
-
-	// FIXME currently ignored
-	// enabledGroupsFlag = flag.String("groups.enabled", "asserts,durability,background_flushing,connections,extra_info,global_lock,index_counters,network,op_counters,op_counters_repl,memory,locks,metrics", "Comma-separated list of groups to use, for more info see: docs.mongodb.org/manual/reference/command/serverStatus/")
-	enabledGroupsFlag = kingpin.Flag("groups.enabled", "Currently ignored").Default("").String()
+	testF = kingpin.Flag("test", "Check MongoDB connection, print buildInfo() information and exit.").Bool()
 )
 
 func main() {
@@ -86,13 +64,7 @@ func main() {
 	if *testF {
 		buildInfo, err := shared.TestConnection(
 			shared.MongoSessionOpts{
-				URI:                   *uriF,
-				TLSConnection:         *tlsF,
-				TLSCertificateFile:    *tlsCertF,
-				TLSPrivateKeyFile:     *tlsPrivateKeyF,
-				TLSCaFile:             *tlsCAF,
-				TLSHostnameValidation: !(*tlsDisableHostnameValidationF),
-				AuthentificationDB:    *authDB,
+				URI: *uriF,
 			},
 		)
 		if err != nil {
@@ -111,20 +83,11 @@ func main() {
 	programCollector := version.NewCollector(program)
 	mongodbCollector := collector.NewMongodbCollector(&collector.MongodbCollectorOpts{
 		URI:                      *uriF,
-		TLSConnection:            *tlsF,
-		TLSCertificateFile:       *tlsCertF,
-		TLSPrivateKeyFile:        *tlsPrivateKeyF,
-		TLSCaFile:                *tlsCAF,
-		TLSHostnameValidation:    !(*tlsDisableHostnameValidationF),
-		DBPoolLimit:              *maxConnectionsF,
 		CollectDatabaseMetrics:   *collectDatabaseF,
 		CollectCollectionMetrics: *collectCollectionF,
 		CollectTopMetrics:        *collectTopF,
 		CollectIndexUsageStats:   *collectIndexUsageF,
 		CollectConnPoolStats:     *mongodbCollectConnPoolStatsF,
-		SocketTimeout:            *socketTimeoutF,
-		SyncTimeout:              *syncTimeoutF,
-		AuthentificationDB:       *authDB,
 	})
 	prometheus.MustRegister(programCollector, mongodbCollector)